requirements.txt
python-dotenv==1.0.1
pip==25.0.1
python-telegram-bot==21.10
requests==2.32.3
openai==1.61.1
pytest~=8.3.4